WebsocketClientBehavior, overridable behaviors
This introduced a new class that contains methods to be called for each connection. Before this modification all the connections should use the same handlers, furthermore the client was described as a dictionary, not so good to keep connection state (context).
Now it is possible to implement pythonic code with client centered objects.